Moved HasNpu method to AppUtils.
