+++ AIDevGallery/NativeMethods.txt
@@ -2,19 +2,3 @@
 DXGI.*
 IDXGIFactory2
 DXGI_ERROR_NOT_FOUND
-
-// DMLHelpers.cs
-IDXCoreAdapterFactory
-IDXCoreAdapter
-IDXCoreAdapterList
-D3D_FEATURE_LEVEL
-DXCoreCreateAdapterFactory
-S_OK
-IDMLDevice
-ID3D12CommandQueue
-ID3D12Device
-D3D12CreateDevice
-DMLCreateDevice
-D3D12_FEATURE_DATA_FEATURE_LEVELS
-D3D12_COMMAND_QUEUE_PRIORITY
-DXCORE_ADAPTER_ATTRIBUTE_D3D12_CORE_COMPUTE
