+++ AIDevGallery/Pages/APIs/APISelectionPage.xaml.cs
@@ -6,7 +6,6 @@ using AIDevGallery.Samples;
 using AIDevGallery.Telemetry.Events;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Navigation;
-using System;
 using System.Collections.Generic;
 using System.Linq;
